Exclude test on iOS/tvOS 9.x

+++ Tests/SQLiteTests/ConnectionTests.swift
@@ -384,6 +384,9 @@ class ConnectionTests : SQLiteTestCase {
     }
 
     func test_concurrent_access_single_connection() {
+        // test can fail on iOS/tvOS 9.x: SQLite compile-time differences?
+        guard #available(iOS 10.0, OSX 10.10, tvOS 10.0, watchOS 2.2, *) else  { return }
+
         let conn = try! Connection("\(NSTemporaryDirectory())/\(UUID().uuidString)")
         try! conn.execute("DROP TABLE IF EXISTS test; CREATE TABLE test(value);")
         try! conn.run("INSERT INTO test(value) VALUES(?)", 0)
